requirements.txt
flet>=0.21
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
